APPS-581 always unwrap optional (#137)
The issue in APPS-581 is that the evaluation of an "at" expression (e.g. array[0]) is expecting a non-optional value. V_Optional doesn't actually appear to be used for anything anymore - we should consider it for removal but I don't want to do so until we have more extensive tests. A short term fix is to automatically unwrap all evaluation results.