Release 0.14.2 (#160)
* add snapshot workflow

* add snapshot workflow

* set version to snapshot, add developer docs

* add test

* fix release notes

* add code generation example

* add test

* APPS-573 escape strings when generating/formatting code unless they are in the command block (#134)

Merge adjacent ExprStrings to make escaping easier, and call escape on strings in the code formatter and generator, as long as they are not in the command section.

* fix snapshot workflow

* improve build.sbt

* fix repository name

* add test

* add comment

* Do not treat placeholder options as occurring 'inPlaceholder' when generating code (#135)

Expressions in placeholders are handled a bit differently, but this should only apply to the expression itself, not any expressions used in placeholder options.

* merge developer docs, add instructions to run make

* APPS-581 always unwrap optional (#137)

The issue in APPS-581 is that the evaluation of an "at" expression (e.g. array[0]) is expecting a non-optional value. V_Optional doesn't actually appear to be used for anything anymore - we should consider it for removal but I don't want to do so until we have more extensive tests. A short term fix is to automatically unwrap all evaluation results.

* APPS-579 Add regression tests for expression evaluation: boolean, substitution. (#136)

* update to dxCommon 0.3.0

* update release notes

* fix quoting of placeholder options (#139)

The previous fix for placeholders was not quite correct - it failed to put quotes around literal values. It also assumed you can't have nested placeholders, which is incorrect. This fixes both issues.

* update versions and changelogs

* use WdlOptions for parsing (#143)

dd Any-type array argument type as valid for concatenation function

* Exclude `loc` field from `equals` and `hashCode` in AST case classes (#144)

Moves loc to the second parameter section in AST case classes. Renames some AST field names for greater clarity. Cleans up parser code and improves error messages. Bumps version.

* Don't strip indent from command block in code formatter/generator (#145)

Previously the code formatter/generator tried to be clever and reformat the code block to use indenting consistent with the rest of the WDL file, but this can break some complex bash scripts. This PR removes that logic and just leaves the command block as-is.

* Store quoting type in AST (#146)

* Store quote type in AST
* Allows any type of expression to be used as the value of the default placeholder option
* No longer performs parameter_meta validation at parse time - it is already done at type-check time

* APPS-589 State machine for formatting expressions (#147)

Implements a simple state machine for formatting WDL expressions, in order to better handle nested quotes and placeholders.

* merge changelog, bump version

* Fix quoting and escaping of import URIs by code generator/formatter (#150)

Fix quoting and escaping of import URIs by code generator/formatter

* add test

* Disable line wrapping in placeholders (#151)

disables line wrapping in placeholders.

* update changelog and version

* Add tool to automatically fix some non-standard WDL sytanx (#149)

Adds tool to automatically fix some non-standard WDL syntax. Improves Eval error messages.

* update version and changelog

* Fix some issues with the `fix` command (#154)

* fix output paths for imports that are relative to a common ancestor; move document writing to separate function shared by Fix, Format, and Upgrade
* allow auto-coercion of read_* return values
* handle case where there's only one file
* update docs

* update makefile

* support coercion from optional type to string within placeholder (#156)

* update docs

* fix snapshot query

* update changelog and version

* update dxCommon version

* add test

* update changelog, add test

* add aliases for v1.1

* Support `container` as alias for `docker` in WDL 1.1 (#158)

* add aliases for v1.1

* support 'container' as an alias for 'docker' in WDL 1.1

* Update Runtime.scala

* Add corpora of public WDLs for testing (#159)

* Allow primitive -> String coercion for WDL versions <= 1.0
* Fixes the grammar for expressions, because the order of operations was not being respected for unary operations
* A parameter_meta key that does not correspond to an input or output parameter is now ignored with a warning unless parsing in strict mode
* Previously, imports were cached using their relative path, which could lead to unexpected results if two imports have the same relative path. Now absolute paths are used.
* Fixes case where the fix tool would generate an invalid declaration

* update dxCommon to published version

* update changelog and version

* update dependency version

Co-authored-by: Riva Nathans <rnathans-cf@dnanexus.com>